CHET-455: [core] move migration stack cleanup service into cleanup package
